feat: polish xmtp
