+++ site/docs/pages/xmtpkit/introduction.mdx
@@ -1,0 +1,8 @@
+---
+title: Introduction to XMTP Kit · OnchainKit
+deescription: Introduction to XMTP Kit
+---
+
+# Introduction to XMTP Kit
+
+XMTP Kit is designed to be modular and flexible, allowing developers to use only the parts of the protocol that they need.
