+++ src/xmtp/getXmtpFrameMessage.test.ts
@@ -1,4 +1,4 @@
-import { getXmtpFrameMessage, isXmtpFrameRequest } from './validation';
+import { getXmtpFrameMessage } from './getXmtpFrameMessage';
 
 const FIXTURES = {
   valid: {
@@ -56,44 +56,5 @@ describe('xmtp validation', () => {
     const invalidResult = await getXmtpFrameMessage(FIXTURES.invalid);
     expect(invalidResult.isValid).toBe(false);
     expect(invalidResult.message).toBeUndefined();
-  });
-});
-
-describe('isXmtpFrameRequest', () => {
-  it('should return true for requests with the correct client protocol', () => {
-    expect(
-      isXmtpFrameRequest({
-        clientProtocol: 'xmtp@2024-02-09',
-        untrustedData: {},
-        trustedData: {},
-      }),
-    ).toBe(true);
-  });
-
-  it('should return false for farcaster requests', () => {
-    expect(
-      isXmtpFrameRequest({
-        untrustedData: {},
-        trustedData: {},
-      }),
-    ).toBe(false);
-  });
-
-  it('should return false for other client protocols', () => {
-    expect(
-      isXmtpFrameRequest({
-        clientProtocol: 'lens@v1',
-        untrustedData: {},
-        trustedData: {},
-      }),
-    );
-  });
-
-  it('should return false for malformed requests', () => {
-    expect(
-      isXmtpFrameRequest({
-        clientProtocol: 'xmtp@2024-02-09',
-      }),
-    ).toBe(false);
   });
 });
