+++ src/xmtp/getXmtpFrameMessage.ts
@@ -26,12 +26,3 @@ export async function getXmtpFrameMessage(payload: XmtpOpenFramesRequest) {
     };
   }
 }
-
-export function isXmtpFrameRequest(payload: any): payload is XmtpOpenFramesRequest {
-  return (
-    !!payload &&
-    !!payload.untrustedData &&
-    !!payload.trustedData &&
-    !!payload.clientProtocol?.startsWith('xmtp@')
-  );
-}
