+++ src/xmtp/isXmtpFrameRequest.ts
@@ -1,0 +1,10 @@
+import { XmtpOpenFramesRequest } from '@xmtp/frames-validator';
+
+export function isXmtpFrameRequest(payload: any): payload is XmtpOpenFramesRequest {
+  return (
+    !!payload &&
+    !!payload.untrustedData &&
+    !!payload.trustedData &&
+    !!payload.clientProtocol?.startsWith('xmtp@')
+  );
+}
